requirements.txt
fastapi
uvicorn
pydantic
PyYAML
ipaddress
pydantic-settings>=2.0.0
pytest
pytest-asyncio
httpx
redis